refactor(add block types and cddl tests): verifcation and refactor
